added new test + refactored mock
